pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepscan"
version = "0.1.0"
description = "DeepScan: A flexible and extensible framework for diagnosing LLMs and MLLMs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "datasets>=2.16.0",
]

[project.optional-dependencies]
# Default install - includes common dependencies for most use cases
default = [
    "torch>=2.1",
    "transformers>=4.37",
    "numpy>=1.24",
    "tqdm>=4.66",
]

# Development dependencies
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

# API server dependencies (removed from core for open-source version)
api = [
    "fastapi>=0.110",
    "uvicorn>=0.24",
]

# Model runner dependencies
qwen = [
    "torch>=2.1",
    "transformers>=4.37",
]
glm = [
    "torch>=2.1",
    # GLM-4.5 series may require a newer Transformers build with glm4_moe support.
    "transformers @ git+https://github.com/huggingface/transformers",
]
# Ministral 3 (multimodal) requires the Mistral tokenizer stack from `mistral-common`
# and, per the official model card, a recent Transformers build (often main/v5).
ministral3 = [
    "torch>=2.1",
    "mistral-common>=1.8.6",
    "transformers @ git+https://github.com/huggingface/transformers",
]

# Evaluator dependencies
tellme = [
    "torch>=2.1",
    "transformers>=4.37",
    "pandas>=2.1",
    "opt_einsum>=3.3",
    "tqdm>=4.66",
]
xboundary = [
    "torch>=2.1",
    "transformers>=4.37",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "seaborn>=0.12",
]
mi_peaks = [
    "torch>=2.1",
    "transformers>=4.37",
    "pandas>=2.1",
    "numpy>=1.24",
]

# Convenience extras
all = [
    "torch>=2.1",
    "transformers>=4.37",
    "pandas>=2.1",
    "opt_einsum>=3.3",
    "tqdm>=4.66",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "seaborn>=0.12",
]

[project.urls]
Homepage = "https://github.com/AI45Lab/DeepScan"
Repository = "https://github.com/AI45Lab/DeepScan"
Issues = "https://github.com/AI45Lab/DeepScan/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["deepscan*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

